requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
hyperliquid-python-sdk>=0.21.0
ta>=0.10.0
arch>=6.0.0
scipy>=1.10.0
pytest>=7.0.0
requests>=2.31.0
joblib>=1.3.0
statsmodels>=0.14.0
